Move inLiquid to lib function (#50333)
